Split dependency code into model and view modules and rename Kind to DependencyKind
