+++ src/controllers/krate/publish.rs
@@ -7,7 +7,6 @@ use std::sync::Arc;
 use hex::ToHex;
 use serde_json;
 
-use dependency;
 use git;
 use render;
 use util::{read_fill, read_le_u32};
@@ -16,6 +15,7 @@ use util::{internal, ChainError};
 use controllers::prelude::*;
 use views::{EncodableCrate, EncodableCrateUpload};
 use models::{Badge, Category, Keyword, NewCrate, NewVersion, Rights, User};
+use models::dependency;
 
 /// Handles the `PUT /crates/new` route.
 /// Used by `cargo publish` to publish a new crate or to publish a new version of an
